feat(tracking): new add track functions

+++ src/main/java/hcmute/techshop/Service/Order/OrderServiceImpl.java
@@ -7,6 +7,7 @@ import hcmute.techshop.Entity.Payment.PaymentEntity;
 import hcmute.techshop.Entity.Product.DiscountEntity;
 import hcmute.techshop.Entity.Product.ProductEntity;
 import hcmute.techshop.Entity.Shipping.ShippingMethodEntity;
+import hcmute.techshop.Enum.EventType;
 import hcmute.techshop.Enum.OrderStatus;
 import hcmute.techshop.Enum.PaymentStatus;
 import hcmute.techshop.Model.Order.OrderItemModel;
@@ -18,6 +19,7 @@ import hcmute.techshop.Repository.Order.OrderRepository;
 import hcmute.techshop.Repository.Payment.PaymentRepository;
 import hcmute.techshop.Repository.Product.ProductRepository;
 import hcmute.techshop.Repository.Shipping.ShippingRepository;
+import hcmute.techshop.Service.Tracking.ITrackingService;
 import lombok.RequiredArgsConstructor;
 import org.modelmapper.ModelMapper;
 import org.springframework.data.domain.Page;
@@ -42,6 +44,7 @@ public class OrderServiceImpl implements IOrderService {
     private final DiscountRepository discountRepository;
     private final ProductRepository productRepository;
     private final PaymentRepository paymentRepository;
+    private final ITrackingService trackingService;
 
     @Override
     public PageResponse<OrderModel> getOrders(OrderStatus orderStatus, int page, int size, Authentication auth) {
@@ -151,7 +154,6 @@ public class OrderServiceImpl implements IOrderService {
         payment.setCreatedAt(LocalDateTime.now());
 
         paymentRepository.save(payment);
-
         return modelMapper.map(savedOrder, OrderModel.class);
     }
 
@@ -165,7 +167,11 @@ public class OrderServiceImpl implements IOrderService {
     }
 
     @Override
-    public void cancelOrder(Integer orderId) {
+    public void cancelOrder(Integer orderId, Authentication auth) {
+        if (auth == null) return;
+
+        UserEntity user = (UserEntity) auth.getPrincipal();
+        trackingService.track(user, EventType.CANCEL_ORDER, "cancel_order with id" + orderId.toString());
         orderRepository.deleteById(orderId);
     }
     @Override
